refactor: VerificationInfo 도메인 분리
